refactored matchmaking as sort, not filter
