+++ workers/eloCalculations.js
